Use consistent origin for webview views (#163602)
For #132464
